Add consistency check for ufs versions against their compiled modules
For new HDFS UFS versions, a new enum describing the version needs to be
added to HdfsVersion.java as well as adding a new ufs module to compile
the jar when building the tarball. It is easy to check for the UFS
modules in a tarball by inspecting the lib directory of a tarball, but
it is not easy to check that the corresponding version enum is defined.
This check ensures that each HDFS UFS module has a corresponding
HdfsVersion by checking that the compilation version matches version
enum regex.

pr-link: Alluxio/alluxio#15955
change-id: cid-caa6d93a2d90b9f52f3f219123ec53124be21fde