Fix JVM pause monitor test
### What changes are proposed in this pull request?

JVM pause monitor test can fail sometimes due to data races.

The thread sleep isnt always accurate, so it uses the more precise timer
(the same that the actual pause monitor uses) and it increases the wait
time by one, because the check in the monitor class is exactly for 250,
which would fail if the time was exactly that.

It also ensures the logs are flushed on write.

### Does this PR introduce any user facing changes?

No.

pr-link: Alluxio/alluxio#16007
change-id: cid-e6e4c8e5b29e0dfe7e0568ac80755293392269d1