Fix journal shutdown deadlock
### What changes are proposed in this pull request?

TakeSnapshot in JournalStateMachine calls a Ratis function that needs to
acquire a lock in the RaftServerProxy.

When the raft server is shutting down it takes the lock in
RaftServerProxy then waits for the state machine thread thread to finish
while holding this lock. This causes a deadlock.

This PR fixes this by ensuring takeSnapshot will not take this lock once
shutdown has started.

### Does this PR introduce any user facing changes?

No

pr-link: Alluxio/alluxio#16042
change-id: cid-77658ce6856d45aebdede58fe5745c83b2cadb25