Fix S3 API writing objects yields BucketNotFound 404
Resolves https://github.com/Alluxio/alluxio/issues/15844

There was an outdated `catch (FileNotFoundException e)` which returned a
`BucketNotFound` S3 Error code (HTTP 404). It was outdated because a
check that the bucket path exists was added since that code had been
written. Furthermore, we needed to return an empty list for the
`children` to `ListBucketResults`.

No user facing changes.

pr-link: Alluxio/alluxio#16074
change-id: cid-8ec241e1bc47e3a361ecc47e2362911b30199afc