Determine the primary master address by calling GetNodeState endpoint
### What changes are proposed in this pull request?
1. In PollingMasterInquireClient, replace the current version service
call with a new introduced GetNodeState grpc endpoint which returns if
the node is primary or standby directly, which internally gets the node
state from the PrimarySelector.
2. Replace the PrimarySelector.State with a protobuf model NodeState.

### Why are the changes needed?

Currently PollingMasterInquireClient determines which master is the
primary master by polling each client to see which one responds. This is
a bit hacky and based on the assumption that standby masters don't have
a gRPC server. So we replace the current call with a new call to
GetNodeState endpoint that tells if the node is primary or not directly.

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#16065
change-id: cid-d55416909f1b1e0f1811821d0a41b80332e4eabb