Deflake AlluxioMasterProcessTest
After looking online, it appears that using PowerMockito with
parameterized tests can be unstable (e.g.
[link](https://stackoverflow.com/questions/2454473/junit-parameterized-tests-together-with-powermock-how)).
Since this test does not require PowerMockito, I used Mockito instead.
I also refactored the `startStopTest` so that I would only wait on the
services that it expects to be started. This speeds up the test
execution.

pr-link: Alluxio/alluxio#16109
change-id: cid-75f6e93128fbf636205656a64ed6541b6f55a306