Allow root sync path to use child sync time
### What changes are proposed in this pull request?

This addresses issue https://github.com/Alluxio/alluxio/issues/15265.

Currently during the metadata sync of a directory, if a child has a more
recent sync time than its parent and does not need to be synced then
this child path will not be synced, and the sync of the parent will
return an error. As a result the parent's last sync time will not be
updated.

This change addresses the issue in two ways.
1. If a child directory does not need to be synced, then it will not be
synced. The parent will then update its sync time only to the oldest
sync time of a child that was not synced (or the current clock time if
all children were synced).
2. If a child file does not need to be synced (but its updated state has
already been loaded from the UFS due to the listing of the parent
directory) then the sync is still performed (because no additional UFS
operations are needed, unless ACL is enabled for the UFS, then an
additional UFS call would be needed so the sync is skipped and the time
is calculated as in 1.).

### Does this PR introduce any user facing changes?

No

pr-link: Alluxio/alluxio#16081
change-id: cid-ea7b4d2fcc8b7690ccdbcd24fdc3574568991e9e