Create inode before updating the MountTable
### What changes are proposed in this pull request?

1. Add method `validateMount` in `MountTable`
2. When doing mount in `DefaultFileSystemMaster`, load the directory
metadata first, then modify the `MountTable`

### Why are the changes needed?
One major challenge we face in #15578 is that when the mount point is
added to `MountTable`, the corresponding inodes don't exist yet. That
makes our proposed change in #15578 hard because by using a
`Trie<Inode>` data structure the inode must exist first.

In this change we propose creating the inode before updating the
MountTable, so when we update the MountTable, we can just update the
`MountTableTrie` with the correct Inode.

### Does this PR introduce any user facing changes?
Nope

pr-link: Alluxio/alluxio#15928
change-id: cid-04a21766e911f7fbdbe5310685af013617cfa8b8