Allow dynamic resource pool returning resources using fifo strategy
### What changes are proposed in this pull request?

Allow dynamic resource pool returning resources in the queue in a FIFO
manner.

### Why are the changes needed?

Currently, the resource pool always returns the last returned resource
on acquire(). This creates some imbalance and might cause some issues if
the resource is a client. This changes mitigates the issue.

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#16142
change-id: cid-78aaf4a3e64987875d73ecb2ef3f5125f2837dda