Fix single master embedded journal checkpoint
Currently, when calling `bin/alluxio fsadmin journal checkpoint` on a
cluster with a single embedded journal master, the information about the
new snapshot does not get propagated back to Ratis. This in turn means
that logs are not pruned and the snapshot is ignored on master restart.
This fix makes sure the new snapshot is taken through Ratis so that all
appropriate functions may be executed.

Removed `mIsSnapshotAllowed` from `RaftJournalSystem` as it is
functionally a duplicate of `mIsLeader` in `JournalStateMachine`.

Redo of #16117.

pr-link: Alluxio/alluxio#16202
change-id: cid-1e60909025f6a7fdf87c1fa94f29cb21137d2da8