Fix load sessionId
### What changes are proposed in this pull request?

use Constant for session id of load API. In this case, we can pick up
the temporary block when the master wants to retry failed blocks. This
is a temporary fix only and would be replaced after we implement the new
session design in the worker.
Also add some error logging.
### Why are the changes needed?

bug fix

### Does this PR introduce any user facing changes?
na

pr-link: Alluxio/alluxio#16217
change-id: cid-4240b1997bc3634989b144f9145cdf5d5fc7684a