Use concurrent map for BlockLockManager
### What changes are proposed in this pull request?

Use `IndexedSet` which uses `ConcurrentHashMap` internally to replace
the shared metadata lock to improve performance under heavy lock
contention.

### Why are the changes needed?

When lock contention is heavy, the `sharedMetadataLock` becomes a
bottleneck of the throughput of lock acquisition and releasing,
defeating the purpose of the individual block locks which are meant to
be a fine-grained lock pool.

See also #16022 for details and motivation for this change.

### Does this PR introduce any user facing changes?

No.

pr-link: Alluxio/alluxio#16078
change-id: cid-c02208c88835d417efa6ee0b34f3ceb8fc396d60