Add getUnsafe() interface for PrimarySelector
### What changes are proposed in this pull request?

Replace the lock with volatile for the state variable in PrimarySelector
because the state assignment is atomic and doesn't depend on any
previous values.

Provide a getUnsafe() method to read the variable without adding a lock.

### Why are the changes needed?

Improve the read performance of the primary selector.

### Does this PR introduce any user facing changes?

Please list the user-facing changes introduced by your change, including
N/A

pr-link: Alluxio/alluxio#16196
change-id: cid-26e2cec6bc5f44ab762d240835ce866f13859f78