Refine add mPendingPaths in InodeSyncStream new type
### What changes are proposed in this pull request?
In the current implementation of InodeSyncStream, when synchronizing a
large directory, because mPendingPaths is a queue, the entire
synchronization process is similar to breadth-first traversal. The
ufsStatusCache will be loaded a lot in advance, which will cause a lot
of memory to be wasted, and even jvm will continue to start full gc
because there is no memory.

### Why are the changes needed?
I added a new option for mPendingPaths, which can make it a stack, the
sync process will become a depth-first traversal, because the leaf nodes
can be reached early, the ufsStatusCache can be recycled early, saving a
lot of memory.

### Does this PR introduce any user facing changes?
alluxio.master.metadata.sync.pending.path.type, can be QUEUE(default) or
STACK

pr-link: Alluxio/alluxio#16204
change-id: cid-4ca33d4d44752f3e7f2689ea9c507f022ae56b51