Avoid canceling duplicate metadata sync prefetch job
### What changes are proposed in this pull request?

During a recursive meta data sync, when a child directory is set to be
fetched from the UFS it will be done in a separate thread. Another
thread will then use this perfected data to sync the path. Currently if
this thread starts running before the prefetch is complete, it will
cancel the prefetch and relaunch it. This PR changes this so it will
instead wait for the thread to complete.

pr-link: Alluxio/alluxio#16273
change-id: cid-f952c774c98054270a7c18fc8d852b829a5df325