Enable java 11 in alluxio base image
By adding an argument to specify the java version, users will be able to
launch Alluxio with java 11 in containers.

Only java 8 and java 11 are supported. If users try build with other
versions, the build would fail with error `failed to solve with frontend
dockerfile.v0: failed to create LLB definition: pull access denied,
repository does not exist or may require authorization: server message:
insufficient_scope: authorization failed`. It is not intuitive at all
but there is no way to print an custom error message because the error
happens when it's trying to load the metadata for all images used in the
dockerfile, before any script runs, where we can put our custom error
message.

This is the first step of a series of updates of the images. The
documentation will be updated later.

pr-link: Alluxio/alluxio#16285
change-id: cid-9b2c6a1f77a323036ac30c9684c76f97929a9fdd