Avoid one copy in PathUtils.findLowestCommonAncestor
### What changes are proposed in this pull request?
An ArrayList that will eventually be converted to an Array is used to
hold the common components in `PathUtils.findLowestCommonAncestor()` .
It is better to use Array because the length of ArrayList is fixed in
the method.

### Why are the changes needed?
`ArrayList.toArray()` dose a copy. Using Array can reduce one copy if we
know very well how many elements there are.

### Does this PR introduce any user facing changes?
NA

pr-link: Alluxio/alluxio#15450
change-id: cid-6f9024842a5bf3881317035ff2629a72360a3306