Add block size for paged blocks
### What changes are proposed in this pull request?

Put block size metadata in paged blocks, so that the information can
persist across worker restarts.

### Why are the changes needed?

The block size is vital for various worker operations. However, page
store does not currently support attaching metadata to pages, therefore
this metadata is lost when the worker restarts and all pages are
invalidated. It would be beneficial that the pages are preserved.

### Does this PR introduce any user facing changes?

No.

pr-link: Alluxio/alluxio#16254
change-id: cid-e1f05dd89bca1470e5074aa2e76c4b0c46884dd4