Mark root as needing sync on backup restore
When restoring from a backup, some modifications exist in UFS that are
not present in Alluxio namespace. This PR marks the root (and in turn
all subdirectories) as needed to be synced with UFS after restoring from
a backup. This amortizes the cost of recovering information from UFS
over the course of the subsequent accesses.

pr-link: Alluxio/alluxio#16351
change-id: cid-e9f81b1e83b14ac90e0285de89125960cd1a0bb7