Fix error in snapshot-taking when using large group ids
To store and distribute RocksDB snapshot files, we compress them into a
`.tar.gz` file. This file fails to be written if the group id or user id
being operated under is too large: `group id '__' is too big ( > 2097151
). Use STAR or POSIX extensions to overcome this limit` or `user id '__'
is too big ( > 2097151 ).`
This PR resolves this issue.

pr-link: Alluxio/alluxio#16365
change-id: cid-3e9172cb2a3c9707ab7c25e75a2c7f24a511fcd7