Optimize MasterWorkerInfo memory usage by introducing fastutil Set
### What changes are proposed in this pull request?

Introduce the HashSet from fastutil (https://fastutil.di.unimi.it/) to
replace the current java HashSet in MasterWorkerInfo

### Why are the changes needed?

MasterWorkerInfo persisted in heap, has become a bottleneck for
persisting more metadata in alluxio masters. This is mainly because it
contains a block id set. If there are too many workers and too many
blocks, this set will exhaust the memory.

A Long object in a java HashSet will take about 64 bytes memory.
However, there are some third party collection libraries for primitive
types. By getting rid of the boxing and unboxing, tons of memories can
be saved and we can also achieve some performance boost.

By using the LongOpenHashSet, we cut down the size of the hash map to a
quarter and save 75% memory.

Fastutil is one of the most used java alternatives .

There are a couple of benchmarks shows the collection components
outperformance the java native one and the size of the collection
components are also significantly smaller.

A couple of benchmarks
https://collection-libs-comparison.develar.org/

https://github.com/Speiger/Primitive-Collections-Benchmarks/blob/master/BENCHMARKS.md

I tested it locally too and I got the same result. Memory size is 75%
smaller. Read is 2x faster, write is 3-5x faster depending on the
operations (put/remove) while the performance of getSize() is almost on
par.

### Does this PR introduce any user facing changes?

N/A

pr-link: Alluxio/alluxio#16443
change-id: cid-b482569e70d6b1f365b527dc62a46ef86d285cb8