Disable delegated backup in irrelevant UTs
### What changes are proposed in this pull request?
After we changed delegated backup to enabled by default, a few UTs are
automatically running on delegated backup and that affected the
stability (in some tests the cluster is not fully mocked so delegated
backup might not even work). This change disables delegated backup in
UTs that are irrelevant.

A few tests are updated and the numbers are from my local runs:

JournalBackupIntegrationTest#backupRestoreZk
enabled - 1 out of 10 success
disabled - 10 out of 10 success

JournalBackupIntegrationTest#backupRestoreEmbedded
enabled - 0 out of 10 success
disabled - 0 out of 10 success
// The flakiness in this test is irrelevant to backup but to election

JournalMigrationIntegrationTest#migrate
enabled - 0 out of 10 success
disabled - 10 out of 10 success

pr-link: Alluxio/alluxio#16482
change-id: cid-c4447df2100474c5e50d36275c6a53a59e3698d9