Fix running Alluxio in k8s as root
These set of env are for the dev image to know the user at build time
(default it's alluxio), so that dev image is able to run processes as
alluxio. But these set of env are then also picked up inside
`entrypoint.sh`, and it would think the user is alluxio, and restart the
processes as user alluxio, even though we run the `entrypoint.sh`
initially as `root`.

The dev image only need to know the `ALLUXIO_UID` to know the user.
Removing other three so that `entrypoint.sh` wouldn't think the user is
non-root and run processes as non-root.

pr-link: Alluxio/alluxio#16514
change-id: cid-8c8c7e179ca3d9c3f9e210b4addcd82dc18dd58d