Avoid allocating new page for read in case of MEM Cache on client.
When MEM cache is used on client, we try to acquire an new/empty page from
the cache manager before reading from external UFS. When the external read
request completes, the data is copied to user-provided buffer, and this
page is added back to cache manager. So, this will avoid allocating new
page for every read.

This will improve the read performance when MEM cache is used on client.

Performance Improvement is expected.

Original-author: Huang Hua <huanghua78@msn.com>