Add BLOCK_COUNT in gRPC WorkerInfoField
### What changes are proposed in this pull request?

Enum variant `BLOCK_COUNT` of `WorkerInfoField` in
`GetWorkerReportOptions.java` was introduced in #14126, but its gRPC
counterpart was not updated accordingly.
This PR adds the missing variant, and reorders the fields of the java
enum.

### Why are the changes needed?

The field cannot be retrieved via RPC, only via internal function call,
because gRPC will fail to serialize that java enum variant.

### Does this PR introduce any user facing changes?

No.

pr-link: Alluxio/alluxio#16457
change-id: cid-e9478f69baa064003bfa5e9309963994c5298eb5