Add replica based annotation
We have added 'ReplicaBasedAnnotation' to consider the replica number of
one block in the whole distributed system when doing an eviction.

We do the simulator for a 5-worker distributed system when the download
speed from the cloud is 0.1x the transfer speed between machines.
If we have 50 positions for each worker, 500 blocks in total, and 10000
random visits, the time cost for ReplicaBased is 93.8% of that for LRU.
If each worker has 100 positions, the time cost for ReplicaBased is
86.1% of that for LRU.

pr-link: Alluxio/alluxio#16357
change-id: cid-210abbeb3859a9c1573e0419a36cd1b15b7e193e