Fix client side config using wrong hash
### What changes are proposed in this pull request?
Revert "Remove ClientContext.mClusterConfHash"
Use the clusterConfHash calculated by the master as the hash version of
all the clients.
fix #16491


### Why are the changes needed?
The master uses Configuration.hash() to calculate the hash, while client
uses clusterConf to calculate the hash version, which will cause the
client to judge that the master and its own clusterConfHash are always
unequal, thus updating the client configuration all the time.

pr-link: Alluxio/alluxio#16523
change-id: cid-4d38cd7e33ee62418a4ab11efdbdb1b554751991