Improve job worker health report
### What changes are proposed in this pull request?

Restructure the worker health report class so user of the class can get
the worker health report at the point in time plus it will be easy to
add many more system information to calculate the worker health.
#16502

### Why are the changes needed?
It would be easy to add new system information like memory, io etc.. of
the worker in calculating the worker health. The new structure of the
class provides the worker health report point in time.

### Does this PR introduce any user facing changes?
No

pr-link: Alluxio/alluxio#16503
change-id: cid-f95a9dbf59cc1777fe6f429ebc810cb4c001268e