Run fault tolerant tests as single fork
This should improve test reliability for the fault tolerant tests as it
runs a single test at a time (see ALLUXIO_DOCKER_FORK_COUNT=1). These
tests are costly as they are starting and stopping Alluxio processes
frequently, which may cause the tests to timeout.

It also adds a wait command to the JournalBackupIntegrationTest to
ensure the cluster is ready after being restarted.

pr-link: Alluxio/alluxio#16595
change-id: cid-3e43b9a7a2a67e7cf0965685b653228406d3c24b