[DOCFIX]Modify some inappropriate descriptions in FAQ.md
### What changes are proposed in this pull request?
The purpose of this pr is to try to fix some comments that were out of
place in FAQ.md and make it more complete.

### Why are the changes needed?
This needs to be modified, otherwise it will affect the user's use.

### Does this PR introduce any user facing changes?
The purpose of this pr is to make it easier for users to use alluxio.

pr-link: Alluxio/alluxio#16593
change-id: cid-24082d4b39acce7d7a9cf23b1db4fa2387b4402e