Fix stale buildVersion when downgrade workers
### What changes are proposed in this pull request?

Clear worker.buildVersion when it's not set in proto.

Note: Since we are using protobuf3 to compile proto2, it's OK to omit
`hasField` check.
If the field is not set, the default instance will be returned.

Steps to reproduce:

1. Start a new cluster (master and worker).
2. Stop the new worker.
3. Start a old worker (prior to
7d8ad9b12ab366ef85c05a7cedfa3a60a562de12) using the same config.
4. Check workers in web UI.

### Why are the changes needed?

Previously, if we downgrade a worker to a versions prior to
7d8ad9b12ab366ef85c05a7cedfa3a60a562de12,
the buildVersion is stale.

### Does this PR introduce any user facing changes?

It fixes a bug in which the buildVersion could be stale.

pr-link: Alluxio/alluxio#16618
change-id: cid-6ab8da9f50471b309adc007cfd2fecab1941fdaa