Refactor simple master services out of main master process classes
This PR refactors many of the simple master process components (the web
server, grpc server, and others) out of their respective `MasterProcess`
classes.

### Motivation
- As HA became the default mode of operation for the master process, all
master tasks must be thought with the perspective of cycling between
primary and standby modes and how to transition between these states (as
opposed to simply starting and stopping)
- Master Process tasks (like the web server and grpc server) used to
always function in lockstep (for example, in the past, if the master
process lost primacy it would always terminate both its grpc server and
its web server). As these tasks have evolved, they no longer behave the
same when gaining/losing primacy (for example, the grpc server may be
stopped when losing primacy while the web server may stay up). This
departure in behavior demands that they be handled differently and
separately.

### Benefits
- Simple services can be added and removed at will. This is particularly
useful for testing when one usually requires only a subset of the
master’s features for a test to run to completion. This should speed
up unit and integration tests.
- Services can be tested in isolation to ensure correctness.
- EE behavior can be separated out into its own service and file, making
code sync between EE and OS much less prone to conflicts.
- In the near future it will be easy to implement and enforce an
ordering to said services. For example, if the web server needs to start
before the journal, this ordering will be quite easy to implement and
enforce based off of the implementation of `MasterRegistry`.
- If advantageous, services could be able to start/promote/demote/stop
in parallel from one another instead of sequentially. This could speed
up tests and debugging sessions.


### Implementation
Essential but simple Master Process tasks have been refactored behind
the `SimpleService` interface. This interface defines only 4 functions:
start, promote, demote, stop. These are to be called at the appropriate
time within the Master Process (the appropriate time is left up the the
implementor of the `MasterProcess` class).
Each service defines a factory which can return different implementation
of said service based on configuration. In an effort to keep the code
straightforward and legible, the is the only place where branching
(a.k.a. if/else statements) based on configuration should be used. If
you foresee start/promote/demote/stop having statements like `if
(Configuration.get…` then it should be made into its own separate
service.

### Testing
- I’ve included tests for all the new services to ensure they work as
intended.
- I’ve prepared a corresponding EE PR to resolve sync issues promptly

pr-link: Alluxio/alluxio#16600
change-id: cid-daf65a9b3004cd8d546cc6c56f6752291a523eed